Implementacion de servicios completado

+++ src/main/java/com/playrole/repository/MensajePrivadoRepositoryInterface.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.playrole.model.MensajePrivado;
 
-public interface MensajePrivadoRepositoryInterface extends JpaRepository<MensajePrivado, Long> {
-	List<MensajePrivado> findByEmisorId_UserIdOrderByFechaEnvioDesc(Long userId);
-    List<MensajePrivado> findByReceptorId_UserIdOrderByFechaEnvioDesc(Long userId);
-    List<MensajePrivado> findByReceptorId_UserIdAndEstadoOrderByFechaEnvioDesc(Long userId, int estado);
-    List<MensajePrivado> findByReceptorId_UserIdAndEstadoInOrderByFechaEnvioDesc(Long userId, List<Integer> estados);
+public interface MensajePrivadoRepositoryInterface extends JpaRepository<MensajePrivado, Integer> {
+	List<MensajePrivado> findByEmisorId_UserIdOrderByFechaEnvioDesc(Integer userId);
+    List<MensajePrivado> findByReceptorId_UserIdOrderByFechaEnvioDesc(Integer userId);
+    List<MensajePrivado> findByReceptorId_UserIdAndEstadoOrderByFechaEnvioDesc(Integer userId, int estado);
+    List<MensajePrivado> findByReceptorId_UserIdAndEstadoInOrderByFechaEnvioDesc(Integer userId, List<Integer> estados);
 }
