+++ src/main/java/com/playrole/repository/PerfilPersonajeRepositoryInterface.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.playrole.model.PerfilPersonaje;
 import java.util.List;
 
-public interface PerfilPersonajeRepositoryInterface extends JpaRepository<PerfilPersonaje, Long> {
+public interface PerfilPersonajeRepositoryInterface extends JpaRepository<PerfilPersonaje, Integer> {
 	
-	List<PerfilPersonaje> findByUserId_UserId(Long userId);
+	List<PerfilPersonaje> findByUserId_UserId(Integer userId);
 }
