+++ src/main/java/com/playrole/repository/UsuarioRepositoryInterface.java
@@ -3,6 +3,6 @@ package com.playrole.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.playrole.model.Usuario;
 
-public interface UsuarioRepositoryInterface extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepositoryInterface extends JpaRepository<Usuario, Integer> {
 
 }
