+++ src/main/java/com/playrole/service/IMensajePrivadoService.java
@@ -6,11 +6,11 @@ import com.playrole.model.MensajePrivado;
 
 public interface IMensajePrivadoService {
 	
-	List<MensajePrivado> mensajesRecibidos(Long idUsuario);
-	List<MensajePrivado> mensajesEnviados(Long idUsuario);
-	List<MensajePrivado> obtenerMensajesNoLeidos(Long idUsuario);
-	MensajePrivado enviarMensaje(Long idEmisor, Long idReceptor, String contenido);
-	List<MensajePrivado> findByEstado(Long idUsuario, List<Integer> estados);
-	void marcarComoLeido(Long idMensaje);
-	void eliminarMensaje(Long id);
+	List<MensajePrivado> mensajesRecibidos(Integer idUsuario);
+	List<MensajePrivado> mensajesEnviados(Integer idUsuario);
+	List<MensajePrivado> obtenerMensajesNoLeidos(Integer idUsuario);
+	MensajePrivado enviarMensaje(Integer idEmisor, Integer idReceptor, String contenido);
+	List<MensajePrivado> findByEstado(Integer idUsuario, List<Integer> estados);
+	void marcarComoLeido(Integer idMensaje);
+	void eliminarMensaje(Integer id);
 }
